+++ R/utils.R
@@ -216,10 +216,12 @@ encode_images = function(x) {
   # only process prose lines and not code blocks
   if (length(p <- prose_index(x)) == 0) return(x)
   xp = x[p]
-  # opening and closing tags of images and audio files
+  # opening and closing tags of images and other media
   rs = matrix(c(
     '!\\[.*?\\]\\(', '\\)',
     '<img .*?src\\s*=\\s*"', '".*?/>',
+    '<audio .*?src\\s*=\\s*"', '".*?>',
+    '<video .*?src\\s*=\\s*"', '".*?>',
     '<source .*?src\\s*=\\s*"',  '".*?>',
     '^background-image: url\\("?', '"?\\)'
   ), 2)